Simplify Command*SubscriptionManagers
